slackbot-workout/requirements.txt
requests
python-dotenv
requests_to_curl
